hardfork test: rename graphql -> client, as we'll add non-graphql interactions to it later
