+++ src/app/hardfork_test/src/internal/client/client.go
@@ -1,4 +1,4 @@
-package graphql
+package client
 
 import (
 	"bytes"
